Refactor: Mypage 유틸 함수 및 커스텀 훅 분리
