+++ src/utils/formatDate.ts
@@ -1,0 +1,8 @@
+export const formatDate = (rawDate?: number) => {
+  if (!rawDate) return '날짜 없음';
+  const date = new Date(rawDate);
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}년 ${month}월 ${day}일`;
+};
